feat(nestjs): add welcome notification endpoint
