+++ playground/nestjs/README.md
@@ -4,6 +4,8 @@ This project is a simple example of how to use Novu Framework with NestJS.
 
 ## Quick start
 
+This quickstart assumes you are running this application from the Novu monorepo and have already installed the dependencies.
+
 Copy the `.env.example` file to `.env` and set the correct environment variables.
 
 ```bash
@@ -13,7 +15,7 @@ cp .env.example .env
 Then, run the application:
 
 ```bash
-npm run start
+pnpm start
 ```
 
 Finally, start Novu Studio and follow the CLI instructions to start creating your NestJS notification workflows:
@@ -25,5 +27,5 @@ npx novu@latest dev
 ## Testing
 
 ```bash
-npm run test
+pnpm test
 ```
