+++ playground/nestjs/src/app.controller.ts
@@ -1,12 +1,31 @@
-import { Controller, Get } from '@nestjs/common';
+import { Controller, Get, Param } from '@nestjs/common';
 import { AppService } from './app.service';
+import { NotificationService } from './notification.service';
+import { UserService } from './user.service';
 
 @Controller()
 export class AppController {
-  constructor(private readonly appService: AppService) {}
+  constructor(
+    private readonly appService: AppService,
+    private readonly notificationService: NotificationService,
+    private readonly userService: UserService,
+  ) {}
 
   @Get()
   getHello(): string {
     return this.appService.getHello();
+  }
+
+  @Get('/welcome/:userId')
+  public async sendWelcomeNotification(@Param('userId') userId: string) {
+    const user = this.userService.getUser(userId);
+
+    return this.notificationService.welcomeWorkflow().trigger({
+      payload: { userId },
+      to: {
+        subscriberId: userId,
+        email: user.email,
+      },
+    });
   }
 }
